refactored main screen setting to add settings tab
